+++ templates/parser.h
@@ -4,8 +4,8 @@
 #include <pybind11/pybind11.h>
 #include <pybind11/numpy.h>
 
-#include "kernels.h"
-#include "exceptions.h"
+#include "george/kernels.h"
+#include "george/exceptions.h"
 
 namespace george {
 
@@ -24,7 +24,7 @@ kernels::Kernel* parse_kernel_spec (py::object& kernel_spec) {
     k1 = parse_kernel_spec(spec1);
     k2 = parse_kernel_spec(spec2);
     if (k1->get_ndim() != k2->get_ndim()) throw dimension_mismatch();
-    int op = py::int_(kernel_spec.attr("operator_type"));
+    size_t op = py::int_(kernel_spec.attr("operator_type"));
     if (op == 0) {
       return new kernels::Sum(k1, k2);
     } else if (op == 1) {
@@ -36,14 +36,14 @@ kernels::Kernel* parse_kernel_spec (py::object& kernel_spec) {
 
 
   kernels::Kernel* kernel;
-  int kernel_type = py::int_(kernel_spec.attr("kernel_type"));
+  size_t kernel_type = py::int_(kernel_spec.attr("kernel_type"));
   switch (kernel_type) {
     {% for spec in specs %}
     case {{ spec.index }}: {
       {% if spec.stationary %}
       py::object metric = kernel_spec.attr("metric");
-      int metric_type = py::int_(metric.attr("metric_type"));
-      int ndim = py::int_(metric.attr("ndim"));
+      size_t metric_type = py::int_(metric.attr("metric_type"));
+      size_t ndim = py::int_(metric.attr("ndim"));
       py::list axes = py::list(metric.attr("axes"));
       bool blocked = py::bool_(kernel_spec.attr("blocked"));
       py::array_t<double> min_block = py::array_t<double>(kernel_spec.attr("min_block"));
@@ -105,7 +105,7 @@ kernels::Kernel* parse_kernel_spec (py::object& kernel_spec) {
       }
 
       {% else %}
-      int ndim = py::int_(kernel_spec.attr("ndim"));
+      size_t ndim = py::int_(kernel_spec.attr("ndim"));
       py::list axes = py::list(kernel_spec.attr("axes"));
       kernel = new kernels::{{ spec.name }} (
           {% for param in spec.params %}
